channelName added
